# api/openapi.yaml
openapi: 3.1.0
info:
  title: Credential Service API
  version: v1
servers:
  - url: https://api.example.com
paths:
...
  /v1/gateway/authorize:
    post:
      summary: Gateway authorization helper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatewayAuthorizeRequest'
      responses:
        '200':
          description: Authorization decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayAuthorizeResponse'
        '403':
          description: Authorization denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayAuthorizeResponse'
components:
  schemas:
...
    GatewayAuthorizeRequest:
      type: object
      properties:
        credential:
          type: string
        credentials:
          type: array
          items:
            type: string
        expected_audience:
          type: string
        want_synthetic_jwt:
          type: boolean
    GatewayAuthorizeResponse:
      type: object
      properties:
        allowed:
          type: boolean
        subject:
          type: string
        acting_on_behalf_of:
          type: string
        delegation_depth:
          type: integer
        claims:
          type: object
        reason:
          type: string
        synthetic_jwt:
          type: string
        api_version:
          type: string
      required: [allowed, api_version]
...